requirements.txt
streamlit
PyPDF2
numpy
faiss-cpu
langchain-google-genai
google-generativeai
requests
typing-extensions
